+++ test/dummy/app/helpers/application_helper.rb
@@ -1,2 +1,5 @@
 module ApplicationHelper
+  def queue_dashboard_path
+    "/queue-dashboard"
+  end
 end