+++ test/dummy/app/jobs/test_job.rb
@@ -2,8 +2,12 @@ class TestJob < ApplicationJob
   queue_as :default
 
   def perform(*args)
-    # Simulate some work
-    sleep 1
-    puts "Job performed with args: #{args.inspect}"
+    # Simulate longer work
+    sleep_time = 15
+    Rails.logger.info "TestJob starting, will sleep for #{sleep_time} seconds"
+    
+    sleep sleep_time
+    
+    Rails.logger.info "TestJob completed with args: #{args.inspect}"
   end
 end